Feat: change script.js to shopUI.js
