+++ index.html
@@ -10,7 +10,7 @@
     />
     <link rel="stylesheet" href="style.css" />
     <link rel="stylesheet" href="src/components/pets/pets.css">
-    <script type="module" src="script.js" defer></script>
+    <script type="module" src="shopUI.js" defer></script>
     <script src="src/components/pets/pets.js" defer></script>
   </head>
   <body>
